+++ packages/@expo/cli/src/prebuildPatch/index.ts
@@ -4,7 +4,7 @@ import chalk from 'chalk';
 import { Command } from '../../bin/cli';
 import { assertArgs, getProjectRoot, printHelp } from '../utils/args';
 
-export const expoUneject: Command = async (argv) => {
+export const expoPrebuildPatch: Command = async (argv) => {
   const args = assertArgs(
     {
       // Types
@@ -22,7 +22,7 @@ export const expoUneject: Command = async (argv) => {
   if (args['--help']) {
     printHelp(
       `Convert native iOS and Android project files into CNG patch files`,
-      chalk`npx expo uneject {dim <dir>}`,
+      chalk`npx expo prebuild:patch {dim <dir>}`,
       [
         chalk`<dir>                                    Directory of the Expo project. {dim Default: Current working directory}`,
         `--no-clean                               Skip cleaning native platform directories`,
@@ -33,22 +33,22 @@ export const expoUneject: Command = async (argv) => {
     );
   }
 
-  // Load modules after the help prompt so `npx expo uneject -h` shows as fast as possible.
+  // Load modules after the help prompt so `npx expo prebuild:patch -h` shows as fast as possible.
   const [
-    // ./unejectAsync
-    { unejectAsync },
+    // ./prebuildPatchAsync
+    { prebuildPatchAsync },
     // ./resolveOptions
     { resolvePlatformOption },
     // ../utils/errors
     { logCmdError },
   ] = await Promise.all([
-    import('./unejectAsync.js'),
+    import('./prebuildPatchAsync.js'),
     import('../prebuild/resolveOptions.js'),
     import('../utils/errors.js'),
   ]);
 
   return (() => {
-    return unejectAsync(getProjectRoot(args), {
+    return prebuildPatchAsync(getProjectRoot(args), {
       // Parsed options
       clean: !args['--no-clean'],
       platforms: resolvePlatformOption(args['--platform']),